Add 104fibo

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -25,7 +25,7 @@ int main(void)
 		n2 = fibn;
 		term++;
 	};
-	printf("%ld%c", sum, '\n');
+	printf("%lld%c", (long long)(sum), '\n');
 	return (0);
 }
